requirements.txt
gradio
gradio_client
python-dotenv
piper-tts
edge_tts
requests
aiohttp
